transfered the letter browsing over to the help page instead of the tutorial
