+++ Project-Folder/SSS/app/src/main/res/layout/activity_help.xml
@@ -6,40 +6,40 @@
     android:layout_height="match_parent"
     tools:context="coen390.nicholas.sss.help">
 
+    <LinearLayout
+        android:layout_width="120dp"
+        android:layout_height="0dp"
+        tools:layout_editor_absoluteX="2dp"
+        tools:layout_editor_absoluteY="0dp"
+        android:id="@+id/linearLayout">
+
+        <ListView
+            android:id="@+id/letterList"
+            android:layout_width="120dp"
+            android:layout_height="512dp"
+            android:layout_marginBottom="0dp"
+            android:layout_marginEnd="285dp"
+            android:layout_marginRight="291dp"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintRight_toRightOf="parent" />
+    </LinearLayout>
+
     <TextView
-        android:id="@+id/textTitleH"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
+        android:id="@+id/showHelpLetter"
+        android:layout_width="110dp"
+        android:layout_height="85dp"
         android:layout_marginBottom="8dp"
         android:layout_marginLeft="8dp"
         android:layout_marginRight="8dp"
         android:layout_marginTop="8dp"
-        android:fontFamily="casual"
-        android:text="LIST OF TRANSLATIONS"
-        android:textAllCaps="false"
-        android:textColor="@android:color/black"
-        android:textSize="24sp"
+        android:textAlignment="center"
+        android:textSize="50sp"
         android:textStyle="bold"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintHorizontal_bias="0.507"
-        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintLeft_toRightOf="@+id/linearLayout"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.074" />
-
-    <ImageView
-        android:id="@+id/imageView"
-        android:layout_width="360dp"
-        android:layout_height="383dp"
-        app:srcCompat="@drawable/signlanguage"
-        app:layout_constraintBottom_toBottomOf="parent"
-        android:layout_marginBottom="8dp"
-        android:layout_marginRight="8dp"
-        app:layout_constraintRight_toRightOf="parent"
-        android:layout_marginLeft="8dp"
-        app:layout_constraintLeft_toLeftOf="parent"
-        android:layout_marginTop="8dp"
-        app:layout_constraintTop_toBottomOf="@+id/textTitleW"
-        app:layout_constraintVertical_bias="0.518" />
+        app:layout_constraintVertical_bias="0.241"
+        app:layout_constraintHorizontal_bias="0.485" />
 
 </android.support.constraint.ConstraintLayout>
