+++ Project-Folder/SSS/app/src/main/res/layout/activity_tutorial_activitiy.xml
@@ -6,40 +6,5 @@
     android:layout_height="match_parent"
     tools:context="coen390.nicholas.sss.tutorialActivitiy">
 
-    <LinearLayout
-        android:layout_width="120dp"
-        android:layout_height="0dp"
-        tools:layout_editor_absoluteX="2dp"
-        tools:layout_editor_absoluteY="0dp"
-        android:id="@+id/linearLayout">
-
-        <ListView
-            android:id="@+id/letterList"
-            android:layout_width="120dp"
-            android:layout_height="512dp"
-            android:layout_marginBottom="0dp"
-            android:layout_marginEnd="285dp"
-            android:layout_marginRight="291dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintRight_toRightOf="parent" />
-    </LinearLayout>
-
-    <TextView
-        android:id="@+id/showHelpLetter"
-        android:layout_width="110dp"
-        android:layout_height="85dp"
-        android:layout_marginBottom="8dp"
-        android:layout_marginLeft="8dp"
-        android:layout_marginRight="8dp"
-        android:layout_marginTop="8dp"
-        android:textAlignment="center"
-        android:textSize="50sp"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toRightOf="@+id/linearLayout"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.241"
-        app:layout_constraintHorizontal_bias="0.485" />
 
 </android.support.constraint.ConstraintLayout>
